Add the DAG and make it simple and functional to send an email
Adding a dag with task and dag structure. By using the function
from the sendEmail program it can send an email. This DAG can
now be run via docker if you add your email as recipient_email.

Use docker compose up to start airflow. Go to localhost:8080

Search for your dag in the search bar. It is called Email_me_info
Once you run the program it should send the recipient and email
with a random number from 1-10 and some other text.