Add requirements to a dockerfile and build dockerfile via compose
Make changes in 1 existing file and add 2 new ones. Required build .
in the docker file. This means that docker compose up will run the
dockerfile that is on the path. Add a new dockerfile, which also
uses the apache/airflow image, but also pip installs requirements.
Add a requirements.txt file with selenium in it. This ensures the
docker image of airflow has selenium installed and can use this
library in the final program.