Add remote selenium server to be run via docker
In the docker compose file, added service selenium to be run on
port 4444. When running the docker compose, this will add a remote
server, which can be used to scrape remotely. Once the server is
running you can also go to localhost:4444 and make sure that the
service is running as expected.