Add helper file scrapeCSV.py which is to be imported to the DAG
This file shall instansiate a remote driver, which should connect
to the selenium server in order to run selenium from airflow.
The first function instansiateRemoteDriver() is also added. It
defines how to create the remote driver. The disable dev shm usage
flag in the options is there to ensure that the program does not
crash as a consequence of the partition being to small.